.pipeline/requirements-dev.txt
python-hcl2
toml
tomli-w
pyyaml
tomli
ruamel.yaml